replace word2vec hardcode
