+++ examples/netflow/config_example_netflow_nodp.json
@@ -1,9 +1,56 @@
 {
     "global_config": {
         "original_data_file": "../traces/ugr16/raw.csv",
-        "dataset_type": "netflow",
+        "overwrite": true,
+        "dataset_meta_type": "single_event_per_row",
+        "dataset_type": null,
         "n_chunks": 10,
         "dp": false
     },
-    "default": "netflow.json"
+    "default": "netflow.json",
+    "pre_post_processor": {
+        "class": "NetsharePrePostProcessor",
+        "config": {
+            "num_train_samples": 50000,
+            "num_test_samples": 50000,
+            "timestamp": {
+                "column": "ts",
+                "generation": true
+            },
+            "word2vec": {
+                "vec_size": 10,
+                "model_name": "word2vec_vecSize",
+                "annoy_n_trees": 100,
+                "pretrain_model_path": null
+            },
+            "metadata": [
+                {
+                    "column": "srcip",
+                    "type": "integer",
+                    "method": "bit"
+                },
+                {
+                    "column": "dstip",
+                    "type": "integer",
+                    "method": "bit"
+                },
+                {
+                    "column": "srcport",
+                    "type": "integer",
+                    "method": "word2vec_port"
+                },
+                {
+                    "column": "dstport",
+                    "type": "integer",
+                    "method": "word2vec_port"
+                },
+                {
+                    "column": "proto",
+                    "type": "string",
+                    "method": "word2vec_proto"
+                }
+            ],
+            "timeseries": []
+        }
+    }
 }
